Merge dev and beta

+++ src/main/java/tk/bluetree242/discordsrvutils/tickets/listeners/TicketDeleteListener.java
@@ -1,3 +1,25 @@
+/*
+ *  LICENSE
+ *  DiscordSRVUtils
+ *  -------------
+ *  Copyright (C) 2020 - 2021 BlueTree242
+ *  -------------
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public
+ *  License along with this program.  If not, see
+ *  <http://www.gnu.org/licenses/gpl-3.0.html>.
+ *  END
+ */
+
 package tk.bluetree242.discordsrvutils.tickets.listeners;
 
 
@@ -15,7 +37,8 @@ import java.sql.SQLException;
 public class TicketDeleteListener extends ListenerAdapter {
 
     public void onTextChannelDelete(TextChannelDeleteEvent e) {
-            DiscordSRVUtils.get().executeAsync(() -> {
+        if (DiscordSRVUtils.get().getMainConfig().bungee_mode()) return;
+        DiscordSRVUtils.get().executeAsync(() -> {
                 try (Connection conn = DiscordSRVUtils.get().getDatabase()) {
                     PreparedStatement p1 = conn.prepareStatement("DELETE FROM tickets WHERE Channel=?");
                     p1.setLong(1, e.getChannel().getIdLong());
